Removed * imports. Improved code formatting and refactored code.
